Add iof:get_table_features, with sensible formatting
Just dumping the response to the console doesn't make sense, as there's
lots of it, and much of it is almost identical.  Attempt to present the
data in a compact but still fairly complete manner, by identifying
unique values for the various feature lists and summarising them as
footnotes.

It's recommended to run iof:debug(off) before requesting table features,
since otherwise all the data will be written to the console anyway.